[REF] web: move error handling code to other file
