+++ addons/web/views/webclient_templates.xml
@@ -166,6 +166,7 @@
         <script type="text/javascript" src="/web/static/src/js/services/local_storage_service.js"></script>
         <script type="text/javascript" src="/web/static/src/js/services/notification_service.js"></script>
         <script type="text/javascript" src="/web/static/src/js/services/crash_manager.js"></script>
+        <script type="text/javascript" src="/web/static/src/js/core/error_utils.js"></script>
         <script type="text/javascript" src="/web/static/src/js/services/session_storage_service.js"></script>
 
         <script type="text/javascript" src="/web/static/src/js/tools/debug_manager.js"></script>
